Move the rest types to rpc-primitives
