+++ crates/subspace-farmer/src/farming.rs
@@ -1,12 +1,13 @@
 use crate::commitments::Commitments;
 use crate::identity::Identity;
 use crate::plot::Plot;
-use crate::rpc::{ProposedProofOfReplicationResponse, RpcClient, SlotInfo, Solution};
+use crate::rpc::RpcClient;
 use anyhow::Result;
 use futures::{future, future::Either};
 use log::{debug, error, info, trace};
 use std::time::Instant;
 use subspace_core_primitives::{crypto, Salt};
+use subspace_rpc_primitives::{ProposedProofOfReplicationResponse, SlotInfo, Solution};
 
 /// Farming Instance to store the necessary information for the farming operations,
 /// and also a channel to stop/pause the background farming task
