+++ crates/subspace-rpc-primitives/src/lib.rs
@@ -18,6 +18,9 @@
 use hex_buffer_serde::{Hex, HexForm};
 use serde::{Deserialize, Serialize};
 use subspace_core_primitives::objects::BlockObjectMapping;
+use subspace_core_primitives::{Salt, Tag};
+
+pub type SlotNumber = u64;
 
 /// Encoded block with mapping of objects that it contains
 #[derive(Clone, Debug, Serialize, Deserialize)]
@@ -49,4 +52,57 @@ pub struct FarmerMetadata {
     /// This constant defines the seed used for deriving pre-genesis objects that will bootstrap
     /// the history.
     pub pre_genesis_object_seed: Vec<u8>,
+}
+
+/// Information about new slot that just arrived
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct SlotInfo {
+    /// Slot number
+    pub slot_number: SlotNumber,
+    /// Slot challenge
+    pub challenge: [u8; 8],
+    /// Salt
+    pub salt: Salt,
+    /// Salt for the next eon
+    pub next_salt: Option<Salt>,
+    /// Acceptable solution range
+    pub solution_range: u64,
+}
+
+/// Proposed proof of space consisting of solution and farmer's secret key for block signing
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct ProposedProofOfReplicationResponse {
+    /// Slot number
+    pub slot_number: SlotNumber,
+    /// Solution (if present) from farmer's plot corresponding to slot number above
+    pub solution: Option<Solution>,
+    /// Secret key, used for signing blocks on the client node
+    pub secret_key: Vec<u8>,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct Solution {
+    pub public_key: [u8; 32],
+    pub piece_index: u64,
+    pub encoding: Vec<u8>,
+    pub signature: Vec<u8>,
+    pub tag: Tag,
+}
+
+impl Solution {
+    pub fn new(
+        public_key: [u8; 32],
+        piece_index: u64,
+        encoding: Vec<u8>,
+        signature: Vec<u8>,
+        tag: Tag,
+    ) -> Self {
+        Self {
+            public_key,
+            piece_index,
+            encoding,
+            signature,
+            tag,
+        }
+    }
 }
